Moved ticket and ticket details to separate components. Added missing translations.
