+++ src/translations/et.json
@@ -35,8 +35,9 @@
   "back": "Tagasi esilehele",
   "modal.comment": "Kommentaar (valikuline)",
   "modal.approve": "Kinnita",
-  "modal.reject": "Reject",
-  "modal.resolve": "Resolve",
+  "modal.reject": "Tagasi saadetud",
+  "modal.resolve": "Lahendatud",
+  "modal.notified": "Teavitatud",
   "modal.approve.description": "Kinnita kirjeldus",
   "modal.reject.description": "Reject description",
   "modal.resolve.description": "Resolve description",
